+++ jsonpickle/ext/gmpy.py
@@ -3,7 +3,7 @@ try:
 except ImportError:
     gmpy = None
 
-from typing import Any, Dict
+from typing import Any
 
 from ..handlers import BaseHandler, HandlerReturn, register, unregister
 
@@ -11,11 +11,11 @@ __all__ = ["register_handlers", "unregister_handlers"]
 
 
 class GmpyMPZHandler(BaseHandler):
-    def flatten(self, obj: gmpy.mpz, data: Dict[str, Any]) -> HandlerReturn:
+    def flatten(self, obj: gmpy.mpz, data: dict[str, Any]) -> HandlerReturn:
         data["int"] = int(obj)
         return data
 
-    def restore(self, data: Dict[str, Any]) -> gmpy.mpz:
+    def restore(self, data: dict[str, Any]) -> gmpy.mpz:
         return gmpy.mpz(data["int"])