+++ jsonpickle/ext/yaml.py
@@ -3,13 +3,11 @@
 The YAML extension module connects jsonpickle to the PyYAML `yaml` module.
 """
 
-from typing import Optional
-
 from ..backend import JSONBackend
 from ..backend import json as jsonpickle_backend
 
 
-def register(backend: Optional[JSONBackend] = None) -> bool:
+def register(backend: JSONBackend | None = None) -> bool:
     """Register the yaml module with jsonpickle's JSONBackend"""
     if backend is None:
         backend = jsonpickle_backend