+++ CHANGES.rst
@@ -11,7 +11,7 @@ v5.0.0
     * **Breaking Change**: Removed the deprecated functions deprecated in 4.1.0 with commit
       59dda25. Additionally, ``util.is_sequence`` was removed due to having only one caller,
       ``util.is_sequence_subclass``. (+563)
-    * Mypy-compatible typing has been added to the entire jsonpickle public API! (#561) (+563)
+    * Mypy-compatible typing has been added to the entire jsonpickle public API! (#561) (+563) (+603)
     * Fixed bug with pickling subclasses of Exception with keyword-only args. (#564) (+565)
     * Removed jsonpickleJS from the tree. (#568) (+569)
     * Add support for the ``types`` module in ``util.importable_name`` (#574) (+575)